app/requirements.txt
prometheus-fastapi-instrumentator
uvicorn
fastapi
pytesseract
transformers
python-multipart
docqa
tesseract
numpy
opencv-python
pillow
